librarymangment.c
Library management is a project that manages
and preserves electronic book data based on the demands of
students. Both students and library administrators can use the
system to keep track of all the books available in the library. It
allows both the administrator and the student to look for the
desired book